Remove dmsgpty references from visor package
